All three platforms have PlatformDeployer in separate module.
